requirements.txt
redis==4.0.2
alembic==1.7.5
celery==4.2.0
celery-redbeat==2.0.0
Cerberus==1.3.4
decorator==5.1.0
gunicorn==20.1.0
honcho==1.0.1
falcon==3.0.1
ipython==7.16.2
ipython-genutils==0.2.0
SQLAlchemy==1.4.27
simplejson==3.17.6
redis==4.0.2
python-rapidjson==1.5
python-dotenv==0.19.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
psycopg2-binary